Move IS_FIAT_ONRAMP_ENABLED into its own file
