+++ src/app/pages/FiatOnramp/isEnabled.ts
@@ -1,0 +1,1 @@
+export const IS_FIAT_ONRAMP_ENABLED = !!process.env.REACT_APP_E2E_TEST
